+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 2.1, 2024-05-23
 ---------------
 - add support for Armageddon's Blade savefiles (issue #5)
+- add support for Restoration of Erathia savefiles
 - fix Vial of Dragon Blood being labelled as Vial of Dragonblood in Shadow of Death savefiles
 - tweak game description layout in HTML export